requirements.txt
# Crypto Forecast - Requirements
# ================================

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Optional: Hyperparameter optimization
optuna>=3.3.0

# API and networking
requests>=2.31.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model persistence
joblib>=1.3.0

# Statistics
scipy>=1.11.0

# Progress bars (optional)
tqdm>=4.65.0

# Flask Dashboard
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.0
flask-bcrypt>=1.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
